requirements.txt
# AI Mobile AgentX - Reformed Architecture Requirements

# Core AI and OCR
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Async and Performance
asyncio-extensions>=0.1.0
aiofiles>=23.0.0

# Database and Caching
sqlite3

# Android Automation (Optional - for ADB integration)
pure-python-adb>=0.3.0

# Utilities
python-dateutil>=2.8.0
typing-extensions>=4.7.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0